Moved NodeKeyDetails to 'security.model'.| #372
