+++ FlowCrypt/src/main/java/com/flowcrypt/email/util/exception/SavePrivateKeyToDatabaseException.kt
@@ -5,7 +5,7 @@
 
 package com.flowcrypt.email.util.exception
 
-import com.flowcrypt.email.api.retrofit.response.model.node.NodeKeyDetails
+import com.flowcrypt.email.security.model.NodeKeyDetails
 
 /**
  * @author Denis Bondarenko
